+++ code/SpatialSort.cpp
@@ -97,8 +97,8 @@ void SpatialSort::FindPositions( const aiVector3D& pPosition, float pRadius, std
 		return;
 
 	// do a binary search for the minimal distance to start the iteration there
-	unsigned int index = mPositions.size() / 2;
-	unsigned int binaryStepSize = mPositions.size() / 4;
+	unsigned int index = (unsigned int)mPositions.size() / 2;
+	unsigned int binaryStepSize = (unsigned int)mPositions.size() / 4;
 	while( binaryStepSize > 1)
 	{
 		if( mPositions[index].mDistance < minDist)
